Add testing for updating events
Within this process a bug was found since if some invalid parameters are
given to the update Event form, it would not show any errors. For now
the quick fix available is doing a manual check within an if statement
in the respective controller. In future commits it is expectable that
this check is performed within the html form with help of thyme leaf
(idea: use same strategy as the one for checking the add event form
which is fully working). Another test is suggested for when that is
implemented: The case where an older date is given to the event.