Give a more elegant solution to update venue bug
In previous commits it was noticed a bug with updating venues where it
would accept a venue to be updated with an empty name. The previous
solution relied in manual checking for it. The fix in this commit relies
on the spring annotation, presenting a more neat solution. However there
is still the problem where the error messages do not show up.